refactor(quad): turn the module into a package, with no behaviour change
quad.py is a flat 823-line module with no layer split, so there is nowhere for a
backend catalogue to live: the import-linter contract forbids pantr._backend from
importing any pantr subpackage, which puts the catalogue inside the module's own
namespace, and the alternative of appending it to quad.py mixes dispatch policy
with quadrature rules in one file.

Split by port scope rather than by layer: _rules.py holds the seven 1D rules and
their shared private helpers, _lattice.py PointsLattice and its factory,
_rule_nd.py QuadratureRule and the two nd factories. __init__.py keeps the module
docstring verbatim, since the docs build renders it, and __all__ unchanged.

Two things that are not tidiness. PointsLattice.__module__ and
QuadratureRule.__module__ are rebound to the package path, because mpi constructs
both and mpi4py pickles for collective calls, so without it a pickle from the new
layout names pantr.quad._lattice and an older pantr cannot load it. And the five
private names are re-exported, without entering __all__, because a separate
downstream consumer imports pantr's private symbols and this repository's CI
cannot see that breakage.

Verified: the whole suite, both quad test files untouched, ruff, mypy over 244
files, both import-linter contracts, the doctest, a pickle round trip, and that
all five private names still resolve.

One gotcha worth knowing: scikit-build-core's editable install bakes a static
module-to-path map, so converting a single-file module into a package breaks
every import of it until pip install -e is rerun. Nothing tracked changes.
